+++ CHANGELOG.md
@@ -16,7 +16,7 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
-## [1.25.0-dev](https://github.com/argilla-io/argilla-server/compare/v1.24.0...v1.25.0)
+## [1.25.0](https://github.com/argilla-io/argilla-server/compare/v1.24.0...v1.25.0)
 
 >[!IMPORTANT]
 > This version include changes related to the search index. So, a reindex is needed.
@@ -27,6 +27,7 @@ These are the section headers that we use:
 - Added `ARGILLA_ES_MAPPING_TOTAL_FIELDS_LIMIT` to work with large dataset annotation flows. ([#31](https://github.com/argilla-io/argilla-server/pull/31))
 - Added support for update options for label and multi-label selection questions. ([#34](https://github.com/argilla-io/argilla-server/pull/34))
 - Added `REINDEX_DATASETS` quickstart environment variable to reindex datasets and records into the search engine. ([#39](https://github.com/argilla-io/argilla-server/pull/39))
+- Added pseudo-random shuffling of records based on the user's `id`. ([#23](https://github.com/argilla-io/argilla-server/pull/23))
 
 ### Changed